Fix: 회원가입 URL, 구조 수정
